shipit_code_coverage: Split code coverage information by test suite

+++ src/shipit_code_coverage/shipit_code_coverage/taskcluster.py
@@ -7,13 +7,13 @@ queue_base = 'https://queue.taskcluster.net/v1/'
 
 
 def get_last_task():
-    r = requests.get(index_base + 'task/gecko.v2.mozilla-central.latest.firefox.linux64-ccov-opt')  # NOQA
+    r = requests.get(index_base + 'task/gecko.v2.mozilla-central.latest.firefox.linux64-ccov-opt')
     last_task = r.json()
     return last_task['taskId']
 
 
 def get_task(branch, revision):
-    r = requests.get(index_base + 'task/gecko.v2.%s.revision.%s.firefox.linux64-ccov-opt' % (branch, revision))  # NOQA
+    r = requests.get(index_base + 'task/gecko.v2.%s.revision.%s.firefox.linux64-ccov-opt' % (branch, revision))
     task = r.json()
     return task['taskId']
 
@@ -46,10 +46,24 @@ def get_tasks_in_group(group_id):
     return tasks
 
 
-def download_artifact(task_id, artifact):
-    r = requests.get(queue_base + 'task/' + task_id + '/artifacts/' + artifact['name'], stream=True)  # NOQA
+def download_artifact(task_id, suite, artifact):
+    artifact_path = 'ccov-artifacts/' + task_id + '_' + suite + '_' + os.path.basename(artifact['name'])
 
-    artifact_path = 'ccov-artifacts/' + task_id + '_' + os.path.basename(artifact['name'])  # NOQA
+    if os.path.exists(artifact_path):
+        return
+
+    r = requests.get(queue_base + 'task/' + task_id + '/artifacts/' + artifact['name'], stream=True)
+
     with open(artifact_path, 'wb') as f:
         r.raw.decode_content = True
         shutil.copyfileobj(r.raw, f)
+
+
+def is_coverage_task(task):
+    return task['task']['metadata']['name'].startswith('test-linux64-ccov')
+
+
+def get_suite_name(task):
+    name = task['task']['metadata']['name']
+    name = name[len('test-linux64-ccov/opt-'):]
+    return '-'.join([p for p in name.split('-') if p != 'e10s' and not p.isdigit()])
